requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
PyYAML>=6.0.1
ccxt>=4.2.0
websockets>=12.0
aiohttp>=3.9.0
asyncio>=3.4.3
pydantic>=2.5.0
numpy>=1.26.0
pandas>=2.1.0

# Database
asyncpg>=0.29.0
aiosqlite>=0.19.0
redis>=5.0.0
SQLAlchemy>=2.0.0
asyncpg-sa>=0.0.1

# Monitoring & Logging
python-json-logger>=2.0.7
loguru>=0.7.2
prometheus-client>=0.19.0

# Notifications
python-telegram-bot>=20.7
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
faker>=22.0.0

# Utilities
tenacity>=8.2.3
schedule>=1.2.0

# Web UI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# AI - Gemini Integration
google-generativeai>=0.8.0
